+++ ChangeLog.md
@@ -4,6 +4,7 @@
 ## Unreleased
 
 * NEW Add support for a signature key of remote action messages. 
+* FIX Redirection instances must not appear in the count of backuped instances.
 
 
 ## 2.1